Add files via upload
This handy tool is a work in progress! Currently, it is able to pull the top 5 posts from the three music subreddits I spend the most time browsing, and email them directly to my email. The subject of the email is the day of the week followed by the date (Day, MM-DD-YYYY). 

Subsequent updates will simplify this process using classes. 

*stroudfake1@gmail.com is a placeholder*